first working version, without descriptors and without setattr/getattr to moleculeinblender
